requirements.txt
aenum==3.1.11
alembic==1.8.1
appdirs==1.4.4
APScheduler==3.9.1
async-generator==1.10
attrs==22.1.0
azure-core==1.26.1
azure-storage-blob==12.14.1
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
camera==1.3.0
case-conversion==2.1.0
certifi==2022.6.15.1
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
chromedriver==2.24.1
click==8.1.3
clipboard==0.0.4
colorama==0.4.5
colour==0.1.5
cryptography==38.0.3
cursor==1.3.5
decorator==4.4.2
dill==0.3.6
distlib==0.3.6
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
filelock==3.8.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-CKEditor==0.4.6
Flask-Excel==0.0.7
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-SSLify==0.1.5
Flask-WTF==1.0.1
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
greenlet==1.1.3
h11==0.14.0
html-text==0.5.2
idna==3.3
imageio==2.22.4
imageio-ffmpeg==0.4.7
infi.systray==0.1.12
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
lml==0.1.0
lxml==4.9.1
Mako==1.2.2
MarkupSafe==2.1.1
maybe-else==0.2.1
mbstrdecoder==1.1.1
moviepy==1.0.3
msal==1.20.0
msoffcrypto-tool==5.0.0
msrest==0.7.1
numpy==1.23.3
O365==2.0.21
oauthlib==3.2.2
office365==0.3.15
Office365-REST-Python-Client==2.3.15
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.0
parsedatetime==2.6
pathmagic==0.3.14
Pillow==9.3.0
platformdirs==2.5.3
prettierfier==1.0.3
proglog==0.1.10
pycparser==2.21
pydantic==1.10.2
pydub==0.25.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.13.0
pyinstrument==4.4.0
pyiotools==0.3.18
PyJWT==2.5.0
pymiscutils==0.3.14
PyPDF2==2.11.1
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyshorteners==1.0.1
PySocks==1.7.1
pysubtypes==0.3.18
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-magic-bin==0.4.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
readchar==4.0.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.5.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stringcase==1.2.0
tabulate==0.9.0
telebot==0.0.4
texttable==1.6.7
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typepy==1.3.0
types-cryptography==3.3.23.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.16.6
vk-api==11.9.9
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
XlsxWriter==3.0.3
